blueprint refactoring
