+++ server/__init__.py
@@ -30,11 +30,11 @@ def create_app():
 		return User.query.get(int(user_id))
 
 	# blueprint for auth routes in our app
-	from .auth import auth as auth_blueprint
+	from .auth.auth import auth as auth_blueprint
 	app.register_blueprint(auth_blueprint)
 
 	# blueprint for non-auth parts of app
-	from .main import main as main_blueprint
+	from .main.main import main as main_blueprint
 	app.register_blueprint(main_blueprint)
 
 	socketio.init_app(app)
