+++ server/auth/auth.py
@@ -2,10 +2,14 @@ from flask import Blueprint, render_template, redirect, url_for, request, flash
 from flask_login import login_user, logout_user, login_required
 from werkzeug.security import generate_password_hash, check_password_hash
 from flask_login import login_user
-from .models import User
-from . import db
+from server.models import User
+from server import db
 
-auth = Blueprint('auth', __name__)
+auth = Blueprint(
+	'auth', 
+	__name__, 
+	template_folder='templates',
+    static_folder='static')
 
 @auth.route('/login')
 def login():
